ref(composables/useTrades.ts): move types to `components/trade/types`
