+++ components/common/ConnectWallet/WalletAssetTrades.vue
@@ -132,7 +132,7 @@
 
 <script setup lang="ts">
 import { NeoIcon, NeoButton, NeoSkeleton } from '@kodadot1/brick'
-import { TradeType } from '@/composables/useTrades'
+import { TradeType } from '@/components/trade/types'
 import { TRADE_TYPE_TO_PROFILE_TAB_MAP } from '@/components/profile/utils'
 import { formatDistanceToNow } from '@/utils/datetime'
