+++ components/trade/types.ts
@@ -18,3 +18,68 @@ export type MakingOfferItem = {
   metadata: string
   sn: string
 }
+
+export enum TradeStatus {
+  ACTIVE = 'ACTIVE',
+  EXPIRED = 'EXPIRED',
+  WITHDRAWN = 'WITHDRAWN',
+  ACCEPTED = 'ACCEPTED',
+}
+
+export type TradeToken = {
+  id: string
+  name: string
+  sn: string
+  currentOwner: string
+  image: string
+  collection: {
+    id: string
+  }
+  meta: Record<string, unknown>
+}
+
+export type TradeConsidered = {
+  id: string
+  name: string
+  currentOwner: string
+  image: string
+}
+
+export type BaseTrade = {
+  id: string
+  price: string
+  expiration: string
+  blockNumber: string
+  status: TradeStatus
+  caller: string
+  offered: TradeToken
+  desired: TradeToken | null
+  considered: TradeConsidered
+  createdAt: Date
+}
+
+export enum TradeDesiredTokenType {
+  SPECIFIC,
+  ANY_IN_COLLECTION,
+}
+
+export enum TradeType {
+  SWAP = 'swap',
+  OFFER = 'offer',
+}
+
+export type Swap = BaseTrade & {
+  surcharge: string | null
+}
+
+export type Offer = BaseTrade
+
+type Trade = Swap | Offer
+
+export type TradeNftItem<T = Trade> = T & {
+  expirationDate: Date
+  type: TradeType
+  desiredType: TradeDesiredTokenType
+  isAnyTokenInCollectionDesired: boolean
+  targets: string[]
+}
