+++ pages/[prefix]/collection/[id]/swaps.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup lang="ts">
-import { TradeType } from '@/composables/useTrades'
+import { TradeType } from '@/components/trade/types'
 
 definePageMeta({
   layout: 'explore-layout',
